[receiver/dockerstats] For V2 implementation, create sensible enabled defaults  (#14093)
This is not a breaking change unless you are using the non-default V2 scraper which is behind a feature gate.

Currently almost every metric is enabled by default, which was done to demonstrate feature parity with the original V1 scraper which enables almost everything by default.

This changes so most metrics are not enabled by default, and instead a sensible set of defaults are enabled. There are many very detailed memory and block IO metric that 99% of users will never need, and all it does it consume needless MTS.

The change brings the default set of metrics from 63 to 14.